requirements.txt
# 量化交易助手 MCP 服务器依赖
mcp-server>=0.1.0
pandas>=2.0.0
requests>=2.28.0
plotly>=5.14.0
logging>=0.5.1
webbrowser>=0.0.1
typing>=3.7.4
pathlib>=1.0.1